requirements.txt
chardet==5.1.0
matplotlib==3.5.3
nltk==3.7
numpy==1.21.5
opencv_python==4.6.0.66
paddleocr==2.6.1.3
pandas==1.3.5
Pillow==9.5.0
pydantic==1.10.2
reportlab==3.6.12
scikit_learn==1.2.2
scipy==1.7.3
sentence_transformers==2.2.2
tensorflow==2.11.0
Unidecode==1.3.6
